:lock: (bcryptjs) Implemented bcryptjs password encryption.
In this commit, implemented bcryptjs password encryption for author creation and author update functionality. Password
encryption is now applied during author creation, as well as during author update. During author update, the password
provided by the author is compared with the existing password stored in the database. If the passwords match, the encryption
remains unchanged. However, if the author has updated their password, a new encryption is generated using bcryptjs library.
This enhancement ensures that password information is securely stored in the system, providing an additional layer of
security for author accounts.